+++ lib/components/systems/facets/UluFacetsHeaderLayout.vue
@@ -1,9 +1,9 @@
 <template>
-  <div class="UluFacetsHeaderLayout">
-    <div class="UluFacetsHeaderLayout__header">
+  <div class="facets-header-layout">
+    <div class="facets-header-layout__header">
       <slot name="header"></slot>
     </div>
-    <div class="UluFacetsHeaderLayout__main">
+    <div class="facets-header-layout__main">
       <slot name="main"></slot>
     </div>
   </div>
@@ -14,7 +14,7 @@
 </script>
 
 <style lang="scss">
-.UluFacetsHeaderLayout__header {
+.facets-header-layout__header {
   display: flex;
   gap: 1rem;
   align-items: center;