+++ lib/components/systems/facets/UluFacetsResults.vue
@@ -1,14 +1,14 @@
 <template>
-  <div class="UluFacetsResults">
+  <div class="facets-results">
     <transition-group
       v-if="items.length"
       :tag="tag"
       :name="transitionName"
-      class="UluFacetsResults__list"
+      class="facets-results__list"
       :class="classes.list"
     >
       <li
-        class="UluFacetsResults__item"
+        class="facets-results__item"
         :class="classes.item"
         v-for="(item, index) in items"
         :key="item.id || index"
@@ -16,7 +16,7 @@
         <slot name="item" :item="item" :index="index"></slot>
       </li>
     </transition-group>
-    <div v-else class="UluFacetsResults__empty">
+    <div v-else class="facets-results__empty">
       <slot name="empty">
         <p>No matching items found.</p>
       </slot>
@@ -36,7 +36,7 @@
     },
     transitionName: {
       type: String,
-      default: 'UluFacetsFade'
+      default: 'facets-fade'
     },
     classes: {
       type: Object,
@@ -46,18 +46,18 @@
 </script>
 
 <style lang="scss">
-  .UluFacetsResults__list {
+  .facets-results__list {
     list-style: none;
     padding: 0;
   }
 
-  .UluFacetsFade-enter-active,
-  .UluFacetsFade-leave-active {
+  .facets-fade-enter-active,
+  .facets-fade-leave-active {
     transition: opacity 0.25s ease;
   }
 
-  .UluFacetsFade-enter-from,
-  .UluFacetsFade-leave-to {
+  .facets-fade-enter-from,
+  .facets-fade-leave-to {
     opacity: 0;
   }
 </style>